Restructure project to create PyPI package
- slightly cleaner project structure
- setup.py etc.
