+++ setup.cfg
@@ -1,0 +1,3 @@
+[metadata]
+description_file=README.md
+license_files=LICENSE.txt
