+++ src/cli.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+
+from dashcamcleaner.main_cli import run_cli
+
+if __name__ == "__main__":
+    run_cli()
