+++ src/dashcamcleaner/__init__.py
@@ -1,0 +1,8 @@
+"""
+dashcamcleaner.
+
+Censor faces and license plates in videos.
+"""
+
+__version__ = "0.2.1"
+__author__ = 'Thomas Fähse'
