+++ src/gui.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+
+from dashcamcleaner.main import run_app
+
+if __name__ == "__main__":
+    run_app()
